refactoring: move those count_* functions out of config.c
