+++ src/kernel/config.h
@@ -119,18 +119,6 @@ struct param;
 #define GIVE_DEFAULT (GIVE_SELF|GIVE_PEASANTS|GIVE_LUXURIES|GIVE_HERBS|GIVE_GOODS)
     int rule_give(void);
 
-#define COUNT_MONSTERS 0x01
-#define COUNT_MIGRANTS 0x02
-#define COUNT_DEFAULT  0x04
-#define COUNT_ALL      0x07
-#define COUNT_UNITS    0x10
-
-    int count_faction(const struct faction * f, int flags);
-    int count_migrants(const struct faction * f);
-    int count_maxmigrants(const struct faction * f);
-    int count_all(const struct faction * f);
-    int count_units(const struct faction * f);
-
     bool has_limited_skills(const struct unit *u);
     const struct race *findrace(const char *, const struct locale *);
